Revamp language/subtitle selection
- Capitalize the first letter of the shown language
- Remove "Auto" from language and subtitle selection (not very useful and quite confusing)
- Always show the tracks, even when there's only one of them
- Cycle through the language tracks when "l" or "L" is pressed
- Cycle through the subtitle tracks when "s" or "S" is pressed
- Fix a regression in previous commits which prevented loading srt files after the movie was started

To go the extra mile, the following improvements should be considered (and are not included
in this commit):

Gstreamer doesn't show the proper Audio/Subtitle track name.
All we get from it is the language code and the language name.
This is a known issue upstream: https://bugzilla.gnome.org/show_bug.cgi?id=615064
When this is fixed, we should modify xplayer to show the full track name.

When cycling through audio/subtitle tracks, the track name is
currently printed to stdout. This should be shown on top of the movie
itself, via OSD. The OSD currently shipping with Xplayer is extremely
buggy and is only designed to show icons though. I would suggest to ditch
it altogether and to add a text overlay on top of the GST play element.