Fix "Fit window to movie"
Fix deprecated call to gtk_window_resize_to_geometry().

Also don't ignore user when asked to resize to dimensions which don't
fit the screen. First, it's never right to ignore a request without
showing an error or some information to the user. Second, the WM
can take care of this. If the window doesn't fit the screen, it's
likely to be allocated the available space. Third, even if it was
resized to something larger than the screen, so what? that's what
the user asked, so do it... the user can use Alt to drag around.