Removed unnecessary boolean checks.
For the functions checking for double quotes and single quotes, both functions checked the value of single_quot or double_quot twice if the first && statement evaluated to false. The functions were changed to take one branch if one value was true, and the other if the other boolean was set to false. This removes the need for extra boolean checks.

In addition, egg_desktop_file_can_launch function was changed to not need a "found" variable.